Move RbR class to root
